requirements.txt
google-cloud-texttospeech
PyMuPDF
python-docx
pydub
pygame
python-dotenv
